requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=-
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via requests
pytelegrambotapi==4.7.1
    # via -r requirements.in
python-dotenv==0.21.1
    # via -r requirements.in
requests==2.28.2
    # via pytelegrambotapi
sqlalchemy==2.0.3
    # via -r requirements.in
typing-extensions==4.4.0
    # via sqlalchemy
urllib3==1.26.14
    # via requests
